+++ src/pages/entry/TitleEntry/TitleEntry.tsx
@@ -219,7 +219,19 @@ export default function TitleEntry({ page }: TitleEntryProps) {
 
   const titleField = (ipad: boolean = false) => {
     const setTitle = (e: string) => {
-      setProposal({ ...getProposal(), title: e.substring(0, MAX_CHAR) });
+      if (countWords(e) < MAX_WORD || (countWords(e) === MAX_WORD && !/\s$/.test(e))) {
+        setProposal({ ...getProposal(), title: e.substring(0, MAX_CHAR) });
+      }
+    };
+
+    const helperFunction = (title: string) => {
+      const baseHelperText = t('title.helper', {
+        current: countWords(title),
+        max: MAX_WORD
+      });
+      return countWords(title) === MAX_WORD
+        ? `${baseHelperText} (MAX WORD COUNT REACHED)`
+        : baseHelperText;
     };
 
     return (
@@ -234,14 +246,7 @@ export default function TitleEntry({ page }: TitleEntryProps) {
           helpers.validate.validateTextEntry(title, setTitle, setTheErrorText, 'TITLE')
         }
         errorText={validateWordCount(getProposal().title)}
-        helperText={
-          getProposal()?.title?.length > 0
-            ? t('title.helper', {
-                current: countWords(getProposal().title),
-                max: MAX_WORD
-              })
-            : ''
-        }
+        helperText={helperFunction(getProposal().title as string)}
         suffix={<ViewIcon toolTip={t('latex.toolTip')} onClick={handleOpenTitleLatexModal} />}
       />
     );