+++ CHANGELOG.md
@@ -19,6 +19,8 @@ UNRELEASED
 * Added: End-to-End tests for target page, testing pulsar timing beam functionality
 * Changed: Target page documentation, includes updated screenshots
 * Changed: Amended Panel Management so that reviewers will be shown once for each review type they are capable of performing
+* Changed: Restricted title and abstract fields, disabling the entry of more than the max word count
+* Changed: Cypress vesion to 15.5.0
 
 7.0.0
 ----------